pull out UrlDataLoader to its own class
